+++ ksql-engine/src/main/java/io/confluent/ksql/KsqlContext.java
@@ -50,7 +50,7 @@ public class KsqlContext {
    *
    * @param streamsProperties
    */
-  KsqlContext(Map<String, Object> streamsProperties) {
+  public KsqlContext(Map<String, Object> streamsProperties) {
     if (streamsProperties == null) {
       streamsProperties = new HashMap<>();
     }
